+++ code/ui.shell/src/components/Body/index.ts
@@ -1,0 +1,1 @@
+export * from './Body';
