+++ code/ui.shell/src/state/shell.ts
@@ -5,12 +5,12 @@ const SHELL = DEFAULT.STATE.SHELL;
 export function create() {
   const tree = props.observable<t.IShellTreeState>(SHELL.tree);
   const body = props.observable<t.IShellBodyState>(SHELL.body);
-  const sidepanel = props.observable<t.IShellSidepanelState>(SHELL.sidepanel);
+  const sidebar = props.observable<t.IShellSidebarState>(SHELL.sidepanel);
 
   const model: t.IShellState = {
     tree,
     body,
-    sidepanel,
+    sidebar,
   };
   return model;
 }
